+++ numpy/core/multiarray.pyi
@@ -386,3 +386,33 @@ def may_share_memory(
     __b: object,
     max_work: Optional[int] = ...,
 ) -> bool: ...
+
+def asarray(
+    a: object,
+    dtype: DTypeLike = ...,
+    order: _OrderKACF = ...,
+    *,
+    like: ArrayLike = ...
+) -> ndarray: ...
+@overload
+def asanyarray(
+    a: _ArrayType,
+    dtype: None = ...,
+    order: _OrderKACF = ...,
+    *,
+    like: ArrayLike = ...
+) -> _ArrayType: ...
+@overload
+def asanyarray(
+    a: object,
+    dtype: DTypeLike = ...,
+    order: _OrderKACF = ...,
+    *,
+    like: ArrayLike = ...
+) -> ndarray: ...
+def ascontiguousarray(
+    a: object, dtype: DTypeLike = ..., *, like: ArrayLike = ...
+) -> ndarray: ...
+def asfortranarray(
+    a: object, dtype: DTypeLike = ..., *, like: ArrayLike = ...
+) -> ndarray: ...
